big update. need to be more disciplined in future

+++ mnbclient/src/components/dashboard/ListExpanded.js
@@ -3,33 +3,83 @@ import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
 
 import ListItem from './ListItem';
-import InputGroup from '../common/InputGroup';
 
 import {
     getList,
     deleteList,
-    addItem,
     purchaseItem,
-    deleteItem
+    deleteItem,
+    removeUserFromList,
+    clearList
 } from '../../actions/listActions';
+
+import { addItemModal, manageUsersModal } from '../../actions/modalActions';
 
 class ListExpanded extends Component {
     constructor() {
         super();
         this.state = {
             list: null,
-            name: '',
-            quantity: ''
+            listPoll: null,
+            active: []
         };
     }
 
-    onChange = e => {
-        this.setState({ [e.target.name]: e.target.value });
+    deleteList = () => {
+        // Prevent everyone except the list author from deleting the list
+        if (this.props.auth.user.id !== this.state.list.author) return;
+
+        this.props.deleteList(this.state.list._id);
+        this.props.clearList(this.state.list._id);
+        this.props.history.goBack();
     };
 
-    deleteList = () => {
-        this.props.deleteList(this.state.list._id);
+    removeSelf = () => {
+        // Prevent list author from attempting to remove themselves
+        if (this.props.auth.user.id === this.state.list.author) return;
+
+        this.props.removeUserFromList(
+            this.props.auth.user.id,
+            this.state.list._id
+        );
+        this.props.clearList(this.state.list._id);
         this.props.history.goBack();
+    };
+
+    topRightButton = isAuthor => {
+        if (isAuthor) {
+            return (
+                <div className="delete-list" onClick={this.deleteList}>
+                    <img src="/icons/delete.svg" alt="delete list" />
+                </div>
+            );
+        } else {
+            return (
+                <div className="remove-self" onClick={this.removeSelf}>
+                    <img src="/icons/close.svg" alt="leave list" />
+                </div>
+            );
+        }
+    };
+
+    openAddItemModal = () => {
+        this.props.addItemModal(this.state.list._id);
+    };
+
+    openManageUsersModal = () => {
+        this.props.manageUsersModal(
+            this.state.list.contributors,
+            this.state.list._id,
+            this.state.list.author
+        );
+    };
+
+    onItemClick = i => {
+        this.setState({
+            active: this.state.active.map((bool, j) => {
+                return i === j ? !bool : false;
+            })
+        });
     };
 
     addItem = e => {
@@ -42,6 +92,12 @@ class ListExpanded extends Component {
     };
 
     deleteItem = item_id => {
+        const newList = Object.assign({}, this.state.list);
+        const itemIndex = newList.items.map(item => item._id).indexOf(item_id);
+        if (itemIndex >= 0) newList.items.splice(itemIndex, 1);
+        this.setState({
+            list: newList
+        });
         this.props.deleteItem(this.state.list._id, item_id);
     };
 
@@ -49,19 +105,21 @@ class ListExpanded extends Component {
         this.props.purchaseItem(this.state.list._id, item_id);
     };
 
-    componentDidMount() {
+    componentWillMount() {
         const { list_id } = this.props.match.params;
-        const userLists = this.props.lists.userLists;
 
-        if (userLists) {
-            const index = userLists.map(list => list._id).indexOf(list_id);
-            if (index < 0) {
-                this.props.getList(list_id);
-            } else {
-                this.setState({ list: userLists[index] });
-            }
-        } else {
-            this.props.getList(list_id);
+        this.props.getList(list_id);
+        const listPoll = setInterval(
+            () => this.props.getList(list_id),
+            1000 * 30
+        );
+        this.setState({ listPoll: listPoll });
+    }
+
+    componentWillUnmount() {
+        if (this.state.listPoll) {
+            clearInterval(this.state.listPoll);
+            this.setState({ listPoll: null });
         }
     }
 
@@ -72,6 +130,11 @@ class ListExpanded extends Component {
                 .indexOf(this.props.match.params.list_id);
             if (index >= 0) {
                 this.setState({ list: nextProps.lists.userLists[index] });
+                this.setState({
+                    active: nextProps.lists.userLists[index].items.map(
+                        () => false
+                    )
+                });
             }
         }
     }
@@ -95,28 +158,38 @@ class ListExpanded extends Component {
                 date.getFullYear()
             ].join('/');
 
+            const userIsAuthor = list.author === this.props.auth.user.id;
+
             return (
                 <div className="list-expanded-container">
                     <div className="top-bar">
                         <div className="return-button" onClick={this.goBack}>
                             <img src="/icons/back.svg" alt="back" />
                         </div>
-                        <div className="delete-list" onClick={this.deleteList}>
-                            <img src="/icons/close.svg" alt="delete list" />
-                        </div>
+                        {this.topRightButton(userIsAuthor)}
                     </div>
                     <div className="list-expanded">
                         <div className="list-info">
                             <div className="title">{list.name}</div>
                             <div className="contributors">
-                                {list.contributors.map(
-                                    ({ username }, index, arr) => {
-                                        if (index !== arr.length - 1) {
-                                            return username + ', ';
-                                        } else {
-                                            return username.toString();
+                                <div className="contributor-names">
+                                    {list.contributors.map(
+                                        ({ username }, index, arr) => {
+                                            if (index !== arr.length - 1) {
+                                                return username + ', ';
+                                            } else {
+                                                return username.toString();
+                                            }
                                         }
-                                    }
+                                    )}
+                                </div>
+                                {list.author === this.props.auth.user.id && (
+                                    <div
+                                        className="contributor-options"
+                                        onClick={this.openManageUsersModal}
+                                    >
+                                        ...
+                                    </div>
                                 )}
                             </div>
                             <div className="date">{dateString}</div>
@@ -128,43 +201,29 @@ class ListExpanded extends Component {
                             )}
                         </div>
                         <div className="list-items">
-                            {list.items.map(item => {
+                            {list.items.map((item, i) => {
                                 return (
                                     <ListItem
                                         id={item._id}
                                         name={item.name}
                                         quantity={item.quantity}
                                         purchased={item.purchased}
+                                        onClick={this.onItemClick}
                                         onPurchase={this.purchaseItem}
                                         onDelete={this.deleteItem}
+                                        active={this.state.active[i]}
+                                        index={i}
                                         key={item._id}
                                     />
                                 );
                             })}
                         </div>
-                    </div>
-                    <div className="bottom-form">
-                        <form onSubmit={this.addItem}>
-                            <InputGroup
-                                name="quantity"
-                                type="number"
-                                value={this.state.quantity}
-                                onChange={this.onChange}
-                            />
-                            <InputGroup
-                                name="name"
-                                placeholder="item name"
-                                type="text"
-                                value={this.state.name}
-                                onChange={this.onChange}
-                            />
-                            <input
-                                className="submit"
-                                type="submit"
-                                onSubmit={this.addItem}
-                                value="+"
-                            />
-                        </form>
+                        <div
+                            className="add-item-button"
+                            onClick={this.openAddItemModal}
+                        >
+                            + Add new item
+                        </div>
                     </div>
                 </div>
             );
@@ -180,5 +239,14 @@ const mapStateToProps = state => ({
 
 export default connect(
     mapStateToProps,
-    { getList, deleteList, addItem, purchaseItem, deleteItem }
+    {
+        getList,
+        deleteList,
+        purchaseItem,
+        deleteItem,
+        addItemModal,
+        removeUserFromList,
+        clearList,
+        manageUsersModal
+    }
 )(withRouter(ListExpanded));
